+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/chat/ChatManager.java
@@ -1,5 +1,6 @@
 package com.eternalcode.core.feature.chat;
 
+import com.eternalcode.annotations.scan.feature.FeatureDocs;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 
@@ -7,6 +8,10 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.UUID;
 
+@FeatureDocs(
+    name = "ChatManager",
+    description = "It allows you to manage chat"
+)
 public class ChatManager {
 
     private final Cache<UUID, Instant> slowdown;
