+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/chat/ChatManagerController.java
@@ -1,10 +1,7 @@
-package com.eternalcode.core.listener.player;
+package com.eternalcode.core.feature.chat;
 
-import com.eternalcode.core.configuration.implementation.PluginConfiguration;
-import com.eternalcode.core.feature.chat.ChatManager;
 import com.eternalcode.core.notification.NoticeService;
 import com.eternalcode.core.util.DurationUtil;
-import org.bukkit.Server;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -14,20 +11,15 @@ import org.bukkit.event.player.AsyncPlayerChatEvent;
 import java.time.Duration;
 import java.util.UUID;
 
-public class PlayerChatListener implements Listener {
+public class ChatManagerController implements Listener {
 
     private final ChatManager chatManager;
     private final NoticeService noticeService;
-    private final PluginConfiguration config;
-    private final Server server;
 
-    public PlayerChatListener(ChatManager chatManager, NoticeService noticeService, PluginConfiguration config, Server server) {
+    public ChatManagerController(ChatManager chatManager, NoticeService noticeService) {
         this.chatManager = chatManager;
         this.noticeService = noticeService;
-        this.config = config;
-        this.server = server;
     }
-
 
     @EventHandler(ignoreCancelled = true, priority = EventPriority.HIGHEST)
     public void onChatSlowMode(AsyncPlayerChatEvent event) {
@@ -62,19 +54,6 @@ public class PlayerChatListener implements Listener {
     @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
     public void markUseChat(AsyncPlayerChatEvent event) {
         this.chatManager.markUseChat(event.getPlayer().getUniqueId());
-    }
-
-    @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
-    public void playSound(AsyncPlayerChatEvent event) {
-        PluginConfiguration.Sounds sound = this.config.sound;
-
-        if (!sound.enableAfterChatMessage) {
-            return;
-        }
-
-        for (Player online : this.server.getOnlinePlayers()) {
-            online.playSound(online.getLocation(), sound.afterChatMessage, sound.afterChatMessageVolume, sound.afterChatMessagePitch);
-        }
     }
 
 }
