+++ eternalcore-core/src/main/java/com/eternalcode/core/listener/player/PlayerChatSoundListener.java
@@ -1,0 +1,34 @@
+package com.eternalcode.core.listener.player;
+
+import com.eternalcode.core.configuration.implementation.PluginConfiguration;
+import org.bukkit.Server;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.AsyncPlayerChatEvent;
+
+public class PlayerChatSoundListener implements Listener {
+
+    private final PluginConfiguration config;
+    private final Server server;
+
+    public PlayerChatSoundListener(PluginConfiguration config, Server server) {
+        this.config = config;
+        this.server = server;
+    }
+
+    @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
+    public void playSound(AsyncPlayerChatEvent event) {
+        PluginConfiguration.Sounds sound = this.config.sound;
+
+        if (!sound.enableAfterChatMessage) {
+            return;
+        }
+
+        for (Player online : this.server.getOnlinePlayers()) {
+            online.playSound(online.getLocation(), sound.afterChatMessage, sound.afterChatMessageVolume, sound.afterChatMessagePitch);
+        }
+    }
+
+}
