+++ eternalcore-core/src/main/java/com/eternalcode/core/listener/player/PlayerJoinListener.java
@@ -1,9 +1,9 @@
 package com.eternalcode.core.listener.player;
 
+import com.eternalcode.annotations.scan.feature.FeatureDocs;
 import com.eternalcode.core.configuration.implementation.PluginConfiguration;
 import com.eternalcode.core.notification.NoticeService;
 import com.eternalcode.core.util.RandomUtil;
-import org.bukkit.GameMode;
 import org.bukkit.Server;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -17,12 +17,10 @@ public class PlayerJoinListener implements Listener {
     private final NoticeService noticeService;
     private final Server server;
 
-    public PlayerJoinListener(PluginConfiguration config, NoticeService noticeService, Server server) {
-        this.config = config;
-        this.noticeService = noticeService;
-        this.server = server;
-    }
-
+    @FeatureDocs(
+        description = "Send a welcome message to a player when they join the server",
+        name = "Player Join Message"
+    )
     @EventHandler
     public void onPlayerJoin(PlayerJoinEvent event) {
         Player player = event.getPlayer();
@@ -35,14 +33,7 @@ public class PlayerJoinListener implements Listener {
                 .send();
         }
 
-        if (this.config.sound.enabledAfterJoin) {
-            for (Player online : this.server.getOnlinePlayers()) {
-                online.playSound(online.getLocation(), this.config.sound.afterJoin, this.config.sound.afterJoinVolume, this.config.sound.afterJoinPitch);
-            }
-        }
-
-        this.noticeService
-            .create()
+        this.noticeService.create()
             .notice(translation -> translation.event().welcomeTitle())
             .notice(translation -> translation.event().welcomeSubtitle())
             .placeholder("{PLAYER}", player.getName())
@@ -57,5 +48,18 @@ public class PlayerJoinListener implements Listener {
             .onlinePlayers()
             .sendAsync();
 
+    }
+
+    @FeatureDocs(
+        description = "Play a sound after a player joins the server",
+        name = "Player Join Sound"
+    )
+    @EventHandler
+    public void onPlayerJoinSound(PlayerJoinEvent event) {
+        if (this.config.sound.enabledAfterJoin) {
+            for (Player online : this.server.getOnlinePlayers()) {
+                online.playSound(online.getLocation(), this.config.sound.afterJoin, this.config.sound.afterJoinVolume, this.config.sound.afterJoinPitch);
+            }
+        }
     }
 }
